requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
boto3==1.34.0
django-storages==1.14.2
twilio==8.10.0
pyotp==2.9.0
qrcode==7.4.2
phonenumbers==8.13.27
django-allauth==0.57.0
channels==4.0.0
channels-redis==4.1.0
gunicorn==21.2.0
Pillow==10.1.0
django-filter==23.5
drf-yasg==1.21.7
sentry-sdk==1.38.0
django-rate-limit==0.5.0
user-agents==2.2.0
geoip2==4.7.0
django-ipware==6.0.2
django-extensions==3.2.3
django-ckeditor==6.7.0
django-phone-field==1.8.1
django-countries==7.5.1
django-import-export==3.3.1
django-debug-toolbar==4.2.0
django-sslserver==0.22
django-axes==6.0.0
django-otp==1.2.0
django-two-factor-auth==1.15.0
weasyprint==60.1
openpyxl==3.1.2
reportlab==4.1.0